move clustering to ML
