+++ src/openms/include/OpenMS/ML/CLUSTERING/sources.cmake
@@ -1,5 +1,5 @@
 ### the directory name
-set(directory include/OpenMS/COMPARISON/CLUSTERING)
+set(directory include/OpenMS/ML/CLUSTERING)
 
 ### list all header files of the directory here
 set(sources_list_h
@@ -23,7 +23,7 @@ foreach(i ${sources_list_h})
 endforeach(i)
 
 ### source group definition
-source_group("Header Files\\OpenMS\\COMPARISON\\CLUSTERING" FILES ${sources_h})
+source_group("Header Files\\OpenMS\\ML\\CLUSTERING" FILES ${sources_h})
 
 set(OpenMS_sources_h ${OpenMS_sources_h} ${sources_h})
