+++ src/openms/includes.cmake
@@ -41,7 +41,7 @@ include(source/FILTERING/NOISEESTIMATION/sources.cmake)
 include(source/FILTERING/ID/sources.cmake)
 include(source/TRANSFORMATIONS/FEATUREFINDER/sources.cmake)
 include(source/TRANSFORMATIONS/CENTROIDING/sources.cmake)
-include(source/COMPARISON/CLUSTERING/sources.cmake)
+include(source/ML/CLUSTERING/sources.cmake)
 include(source/COMPARISON/SPECTRA/sources.cmake)
 include(source/QC/sources.cmake)
 if(NOT DISABLE_OPENSWATH)
@@ -86,7 +86,7 @@ include(include/OpenMS/ANALYSIS/TARGETED/sources.cmake)
 include(include/OpenMS/ANALYSIS/TOPDOWN/sources.cmake)
 include(include/OpenMS/ANALYSIS/XLMS/sources.cmake)
 
-include(include/OpenMS/COMPARISON/CLUSTERING/sources.cmake)
+include(include/OpenMS/ML/CLUSTERING/sources.cmake)
 include(include/OpenMS/COMPARISON/SPECTRA/sources.cmake)
 include(include/OpenMS/FILTERING/BASELINE/sources.cmake)
 include(include/OpenMS/FILTERING/CALIBRATION/sources.cmake)
