+++ src/openms/source/ML/CLUSTERING/ClusteringGrid.cpp
@@ -7,7 +7,7 @@
 // --------------------------------------------------------------------------
 //
 
-#include <OpenMS/COMPARISON/CLUSTERING/ClusteringGrid.h>
+#include <OpenMS/ML/CLUSTERING/ClusteringGrid.h>
 
 #include <functional>
 #include <sstream>
