+++ src/tests/class_tests/openms/source/HashGrid_test.cpp
@@ -9,7 +9,7 @@
 #include <OpenMS/CONCEPT/ClassTest.h>
 #include <OpenMS/test_config.h>
 
-#include <OpenMS/COMPARISON/CLUSTERING/HashGrid.h>
+#include <OpenMS/ML/CLUSTERING/HashGrid.h>
 
 #include <limits>
 #include <iostream>
